requirements.txt
streamlink>=6.7.4
requests>=2.31.0
python-dotenv>=1.0.0
pathlib2>=2.3.7; python_version < "3.4"